+++ scripts/prepare-submodule.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+. scripts/init-submodule.sh
+. .docker/scripts/reset-submodule.sh
+. scripts/update-submodule.sh